lucetc: refactor so faerie-specific stuff only lives in compiler & output/faerie
