+++ lucetc/src/lib.rs
@@ -4,7 +4,6 @@ mod compiler;
 mod decls;
 mod error;
 mod function;
-mod function_manifest;
 mod heap;
 mod load;
 mod module;
