+++ lucetc/src/table.rs
@@ -7,8 +7,7 @@ use byteorder::{LittleEndian, WriteBytesExt};
 use cranelift_codegen::entity::EntityRef;
 use cranelift_module::{Backend as ClifBackend, DataContext, Module as ClifModule};
 use cranelift_wasm::{TableElementType, TableIndex};
-use faerie::{Artifact, Link};
-use failure::{format_err, Error, ResultExt};
+use failure::{format_err, ResultExt};
 use std::io::Cursor;
 
 /// This symbol will be used to reference the `tables` field in `Module` - a sequence of tables.
@@ -51,18 +50,6 @@ fn table_elements(decl: &TableDecl<'_>) -> Result<Vec<Elem>, LucetcError> {
     }
 
     Ok(elems)
-}
-
-pub fn link_tables(tables: &[Name], obj: &mut Artifact) -> Result<(), Error> {
-    for (idx, table) in tables.iter().enumerate() {
-        obj.link(Link {
-            from: TABLE_SYM,
-            to: table.symbol(),
-            at: (TABLE_REF_SIZE * idx) as u64,
-        })
-        .context(LucetcErrorKind::Table)?;
-    }
-    Ok(())
 }
 
 pub fn write_table_data<B: ClifBackend>(
